+++ src/CleanRegex/Internal/Match/GroupBy/DetailFunction.php
@@ -1,9 +1,0 @@
-<?php
-namespace TRegx\CleanRegex\Internal\Match\GroupBy;
-
-use TRegx\CleanRegex\Match\Details\Detail;
-
-interface DetailFunction
-{
-    public function apply(Detail $detail);
-}